Implement health check endpoint with HealthHandler and HealthResponse

+++ src/main/java/com/sqldataservice/api/feature/anime/detail/DetailAnimeResponse.java
@@ -4,13 +4,13 @@ import com.sqldataservice.api.shared.AnimeSummaryItem;
 import com.sqldataservice.api.shared.KeyValueItem;
 
 public record DetailAnimeResponse(int id, String title, String synopsis, String imageUrl, KeyValueItem type,
-        int episodes,
-        KeyValueItem status, KeyValueItem rating, KeyValueItem[] genres, KeyValueItem[] studios,
-        KeyValueItem[] themes,
-        KeyValueItem[] demographics,
-        KeyValueItem source, KeyValueItem[] licensors, KeyValueItem[] producers,
-        KeyValueItem[] streamingServices,
-        AnimeSummaryItem[] recommendations,
-        DetailAnimeResponseVoice[] voices) {
+		Double episodes,
+		KeyValueItem status, KeyValueItem rating, KeyValueItem[] genres, KeyValueItem[] studios,
+		KeyValueItem[] themes,
+		KeyValueItem[] demographics,
+		KeyValueItem source, KeyValueItem[] licensors, KeyValueItem[] producers,
+		KeyValueItem[] streamingServices,
+		AnimeSummaryItem[] recommendations,
+		DetailAnimeResponseVoice[] voices) {
 
 }
